requirements.txt
azure-functions==1.17.0
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
requests==2.31.0
urllib3==2.0.7